+++ src/components/JobList/JobList.css
@@ -1,0 +1,5 @@
+.job-list {
+  display: flex;
+  list-style-type: none;
+  text-align: center;
+}
