+++ src/components/Profile/Profile.css
@@ -1,0 +1,3 @@
+.profileImage {
+  width: 300px;
+}
