+++ src/components/Profile/index.js
@@ -1,0 +1,1 @@
+export { Profile as default } from './Profile';
