--- src/index.css
@@ -1,9 +1,0 @@
-.profileImage {
-    width: 300px;
-}
-
-.job-list {
-    display: flex;
-    list-style-type: none;
-    text-align: center;
-}
